--- a/paint/fill.go
+++ b/paint/fill.go
@@ -6,11 +6,8 @@ import (
 	"math"
 
 	"github.com/anthonynsimon/bild/clone"
+	"github.com/anthonynsimon/bild/math/f64"
 	"github.com/anthonynsimon/bild/util"
-)
-
-const (
-	maxDistance = 510.0
 )
 
 type fillPoint struct {
@@ -24,8 +21,9 @@ type fillPoint struct {
 // FloodFill fills a area of the image with a provided color and returns the new image.
 // Parameter sp is the starting point of the fill.
 // Parameter c is the fill color.
-// Parameter fuzz is the percentage of maximum color distance tolerated when flooding the area.
-func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *image.RGBA {
+// Parameter t is the tolerance and is of the range 0 to 255. It represents the max amount of
+// difference between colors for them to be considered similar.
+func FloodFill(img image.Image, sp image.Point, c color.Color, t uint8) *image.RGBA {
 
 	var st util.Stack
 	var point fillPoint
@@ -38,7 +36,6 @@ func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *im
 		return im
 	}
 
-	fuzzSquared := math.Pow(maxDistance*fuzz/100, 2)
 	matchColor := color.NRGBAModel.Convert(im.At(sp.X, sp.Y)).(color.NRGBA)
 
 	st.Push(fillPoint{sp.X, sp.Y, true, true, 0, 0})
@@ -62,7 +59,7 @@ func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *im
 					break
 				}
 				pixOffset = im.PixOffset(xpos, point.Y)
-				if isColorMatch(im, pixOffset, matchColor, fuzzSquared) {
+				if isColorMatch(im, pixOffset, matchColor, t) {
 					im.Set(xpos, point.Y, c)
 					visited[pixOffset] = true
 				} else {
@@ -84,7 +81,7 @@ func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *im
 				}
 
 				pixOffset = im.PixOffset(xpos, point.Y)
-				if isColorMatch(im, pixOffset, matchColor, fuzzSquared) {
+				if isColorMatch(im, pixOffset, matchColor, t) {
 					im.Set(xpos, point.Y, c)
 					visited[pixOffset] = true
 				} else {
@@ -110,7 +107,7 @@ func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *im
 					if point.MarkedFromBelow == true || outOfPreviousRange {
 						if point.Y > 0 {
 							pixOffset = im.PixOffset(x, point.Y-1)
-							if false == visited[pixOffset] && isColorMatch(im, pixOffset, matchColor, fuzzSquared) {
+							if !visited[pixOffset] && isColorMatch(im, pixOffset, matchColor, t) {
 								skipCheckBelow = true
 								st.Push(fillPoint{x, (point.Y - 1), true, false, leftFillEdge, rightFillEdge})
 							}
@@ -125,7 +122,7 @@ func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *im
 						if point.Y < maxY {
 
 							pixOffset = im.PixOffset(x, point.Y+1)
-							if false == visited[pixOffset] && isColorMatch(im, pixOffset, matchColor, fuzzSquared) {
+							if !visited[pixOffset] && isColorMatch(im, pixOffset, matchColor, t) {
 								skipCheckAbove = true
 								st.Push(fillPoint{x, (point.Y + 1), false, true, leftFillEdge, rightFillEdge})
 							}
@@ -139,25 +136,30 @@ func FloodFill(img image.Image, sp image.Point, c color.Color, fuzz float64) *im
 	return im
 }
 
-func isColorMatch(im *image.RGBA, pixel int, mc color.NRGBA, fuzzSquared float64) bool {
+func isColorMatch(im *image.RGBA, pos int, mc color.NRGBA, tolerance uint8) bool {
+	c2 := color.RGBA{R: im.Pix[pos+0], G: im.Pix[pos+1], B: im.Pix[pos+2], A: im.Pix[pos+3]}
 
-	i := pixel
-	c1 := mc
-	c2 := color.NRGBA{R: im.Pix[i+0], G: im.Pix[i+1], B: im.Pix[i+2], A: im.Pix[i+3]}
+	rDiff := float64(subuint8(mc.R, c2.R))
+	gDiff := float64(subuint8(mc.G, c2.G))
+	bDiff := float64(subuint8(mc.B, c2.B))
 
-	rDiff := float64(c1.R) - float64(c2.R)
-	gDiff := float64(c1.G) - float64(c2.G)
-	bDiff := float64(c1.B) - float64(c2.B)
-	aDiff := float64(c1.A) - float64(c2.A)
+	distance := math.Sqrt(rDiff*rDiff + gDiff*gDiff + bDiff*bDiff)
 
-	distanceR := math.Max(math.Pow(rDiff, 2), math.Pow(rDiff-aDiff, 2))
-	distanceG := math.Max(math.Pow(gDiff, 2), math.Pow(gDiff-aDiff, 2))
-	distanceB := math.Max(math.Pow(bDiff, 2), math.Pow(bDiff-aDiff, 2))
-	distance := (distanceR + distanceG + distanceB) / 3
+	value := uint8(f64.Clamp(distance, 0, 255))
 
-	if distance > fuzzSquared {
+	if value > tolerance {
 		return false
 	}
 
 	return true
+}
+
+// helper function to avoid uint8 overflow during subtraction
+func subuint8(a, b uint8) uint8 {
+	if b > a {
+		return b - a
+	} else if a > b {
+		return a - b
+	}
+	return 0
 }