… paint/fill_test.go