Corrected range of FloodFill tolerance to more familiar one.

As a consequence the euclidean distance calculation could be simplified and a ~26% performance boost was obtained.

```
benchmark                old ns/op     new ns/op     delta
BenchmarkFloodFill-8     94928744      69608378      -26.67%

benchmark                old allocs     new allocs     delta
BenchmarkFloodFill-8     264417         264419         +0.00%

benchmark                old bytes     new bytes     delta
BenchmarkFloodFill-8     24389554      24390124      +0.00%
```

This closes #31